refactor(migration-sdk-viem): rename files
