+++ packages/migration-sdk-viem/src/positions/supply/aaveV3Optimizer.supply.ts
@@ -27,7 +27,7 @@ import {
 } from "viem";
 import { signTypedData } from "viem/actions";
 import { morphoAaveV3Abi } from "../../abis/aaveV3Optimizer.js";
-import { getMorphoAaveV3ManagerApprovalTypedData } from "../helpers/signatures.js";
+import { getMorphoAaveV3ManagerApprovalTypedData } from "../signature/aaveV3Optimizer.js";
 import {
   type IMigratableSupplyPosition,
   MigratableSupplyPosition,
